requirements.txt
pyserial:3.5
pillow:11.0.0